Renames lib/sshpm/entities.rb to lib/sshpm/host.rb
The file only contained `SSHPM::Host` declaration, therefore it
made no sense to have a name other than `host.rb`
