+++ lib/sshpm/host.rb
